+++ src/utils/index.js
@@ -1,0 +1,4 @@
+import matchLineNumber from './matchLineNumber';
+
+// eslint-disable-next-line import/prefer-default-export
+export { matchLineNumber };